Only execute if configured in the settings
