+++ src/options/options.html
@@ -15,7 +15,7 @@
                 <input id="gitpod-open-as-popup" type="checkbox"/>
             </div>
             <div class="row">
-                <label>Replace <kbd>.</kbd> keybind:</label>
+                <label>Replace <kbd>.</kbd> keybind (opens Gitpod instead of vscode.dev):</label>
                 <input id="gitpod-replace-keybind" type="checkbox"/>
             </div>
             <div class="row">
